Assign elastic ip to an ec2 instance
An Elastic IP address is a static IPv4 address designed for dynamic
cloud computing. An Elastic IP address is associated with your AWS
account. With an Elastic IP address, you can mask the failure of an
instance or software by rapidly remapping the address to another
instance in your account.

An Elastic IP address is a public IPv4 address, which is reachable from
the internet. If your instance does not have a public IPv4 address, you
can associate an Elastic IP address with your instance to enable
communication with the internet; for example, to connect to your
instance from your local computer.